refactor(frontend): 提取 MCP 配置编辑器的共享代码，消除重复
将 mcp-server-setting-button.tsx 和 add-mcp-server-button.tsx 中的
重复代码提取到共享组件 MCPConfigEditor 和常量文件中：

- 创建 MCP_CONFIG_PLACEHOLDER_TEXT 常量，提取配置示例文本
- 创建 MCPConfigEditor 组件，包含：
  - JSON 表单 schema
  - 模式切换逻辑（表单 ↔ JSON）
  - 高级模式表单结构
- 重构两个按钮组件使用共享编辑器
- 更新测试用例匹配新的 tab 名称

减少约 263 行重复代码，提高代码可维护性。

Fixes #3274
