+++ apps/frontend/src/components/mcp-server-setting-button.tsx
@@ -1,33 +1,16 @@
-import { McpServerForm } from "@/components/mcp-server-form";
-import { Button } from "@/components/ui/button";
+import { MCPConfigEditor } from "@/components/mcp-config-editor";
 import {
   Dialog,
-  DialogClose,
   DialogContent,
   DialogDescription,
-  DialogFooter,
   DialogHeader,
   DialogTitle,
   DialogTrigger,
 } from "@/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormField,
-  FormItem,
-  FormMessage,
-} from "@/components/ui/form";
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import { Textarea } from "@/components/ui/textarea";
 import { useNetworkServiceActions } from "@/providers/NetworkServiceProvider";
 import { mcpFormSchema } from "@/schemas/mcp-form";
 import { useConfig } from "@/stores/config";
-import {
-  apiConfigToForm,
-  formToApiConfig,
-  formToJson,
-  jsonToFormData,
-} from "@/utils/mcpFormConverter";
+import { apiConfigToForm, formToApiConfig } from "@/utils/mcpFormConverter";
 import { validateMCPConfig } from "@/utils/mcpValidation";
 import { zodResolver } from "@hookform/resolvers/zod";
 import type { MCPServerConfig } from "@xiaozhi-client/shared-types";
@@ -35,14 +18,7 @@ import { SettingsIcon } from "lucide-react";
 import { useCallback, useState } from "react";
 import { useForm } from "react-hook-form";
 import { toast } from "sonner";
-import z from "zod";
-
-// 高级模式的 JSON 表单 schema
-const jsonFormSchema = z.object({
-  config: z.string().min(2, {
-    message: "配置不能为空",
-  }),
-});
+import type { z } from "zod";
 
 export function McpServerSettingButton({
   mcpServer,
@@ -53,8 +29,6 @@ export function McpServerSettingButton({
 }) {
   const [open, setOpen] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
-  const [inputMode, setInputMode] = useState<"form" | "json">("form");
-  const [jsonInput, setJsonInput] = useState<string>("");
   const config = useConfig();
   const { updateConfig } = useNetworkServiceActions();
 
@@ -67,58 +41,11 @@ export function McpServerSettingButton({
     defaultValues: defaultFormValues,
   });
 
-  // 高级模式的表单实例（保持原有验证逻辑）
-  const advancedForm = useForm<z.infer<typeof jsonFormSchema>>({
-    resolver: zodResolver(jsonFormSchema),
-    defaultValues: {
-      config: JSON.stringify(
-        { mcpServers: { [mcpServerName]: mcpServer } },
-        null,
-        2
-      ),
-    },
-  });
-
-  // 当弹窗关闭时重置状态
-  const handleOpenChange = useCallback(
-    (newOpen: boolean) => {
-      if (!newOpen) {
-        // 重置表单和状态
-        form.reset(defaultFormValues);
-        advancedForm.reset();
-        setJsonInput("");
-        setInputMode("form");
-      }
-      setOpen(newOpen);
-    },
-    [form, advancedForm, defaultFormValues]
-  );
-
-  // 处理模式切换
-  const handleModeChange = useCallback(
-    (newMode: string) => {
-      if (newMode !== "form" && newMode !== "json") {
-        return; // 忽略无效值
-      }
-
-      if (newMode === "json" && inputMode === "form") {
-        // 表单 → JSON
-        const formValues = form.getValues();
-        try {
-          setJsonInput(formToJson(formValues));
-        } catch {
-          setJsonInput("");
-        }
-      } else if (newMode === "form" && inputMode === "json") {
-        // JSON → 表单
-        const formData = jsonToFormData(jsonInput);
-        if (formData) {
-          form.reset(formData);
-        }
-      }
-      setInputMode(newMode);
-    },
-    [inputMode, form, jsonInput]
+  // 初始 JSON 配置字符串
+  const initialJsonConfig = JSON.stringify(
+    { mcpServers: { [mcpServerName]: mcpServer } },
+    null,
+    2
   );
 
   // 表单模式提交处理
@@ -174,8 +101,8 @@ export function McpServerSettingButton({
   );
 
   // 高级模式提交处理
-  const handleAdvancedSubmit = useCallback(
-    async (values: z.infer<typeof jsonFormSchema>) => {
+  const handleJsonSubmit = useCallback(
+    async (values: { config: string }) => {
       if (!config) {
         toast.error("配置数据未加载，请稍后重试");
         return;
@@ -244,6 +171,29 @@ export function McpServerSettingButton({
     [config, mcpServerName, updateConfig]
   );
 
+  // 使用 MCPConfigEditor 组件
+  const editor = MCPConfigEditor({
+    form,
+    defaultValues: defaultFormValues,
+    onFormSubmit: handleFormSubmit,
+    onJsonSubmit: handleJsonSubmit,
+    disabled: isLoading,
+    isLoading,
+    submitText: "保存",
+    initialJsonConfig,
+  });
+
+  // 当弹窗关闭时重置状态
+  const handleOpenChange = useCallback(
+    (newOpen: boolean) => {
+      if (!newOpen) {
+        editor.resetState();
+      }
+      setOpen(newOpen);
+    },
+    [editor]
+  );
+
   return (
     <Dialog open={open} onOpenChange={handleOpenChange}>
       <DialogTrigger asChild>
@@ -264,93 +214,7 @@ export function McpServerSettingButton({
         </DialogHeader>
 
         {/* 模式切换 */}
-        <Tabs value={inputMode} onValueChange={handleModeChange}>
-          <TabsList className="grid w-full grid-cols-2">
-            <TabsTrigger value="form">表单模式</TabsTrigger>
-            <TabsTrigger value="json">高级模式 (JSON)</TabsTrigger>
-          </TabsList>
-
-          {/* 表单模式 */}
-          <TabsContent value="form" className="mt-4">
-            <McpServerForm
-              form={form}
-              defaultValues={defaultFormValues}
-              onSubmit={handleFormSubmit}
-              disabled={isLoading}
-              submitText={isLoading ? "保存中..." : "保存"}
-            />
-          </TabsContent>
-
-          {/* 高级模式 */}
-          <TabsContent value="json" className="mt-4">
-            <Form {...advancedForm}>
-              <form onSubmit={advancedForm.handleSubmit(handleAdvancedSubmit)}>
-                <div className="grid gap-4">
-                  <FormField
-                    control={advancedForm.control}
-                    name="config"
-                    render={({ field }) => (
-                      <FormItem>
-                        <FormControl>
-                          <Textarea
-                            className="resize-none h-[300px] font-mono text-sm"
-                            disabled={isLoading}
-                            placeholder={`支持三种通信方式：
-
-1. 本地进程 (stdio):
-{
-  "mcpServers": {
-    "local-server": {
-      "command": "npx",
-      "args": ["-y", "@example/mcp-server"]
-    }
-  }
-}
-
-2. 服务器推送 (SSE):
-{
-  "mcpServers": {
-    "sse-server": {
-      "type": "sse",
-      "url": "https://example.com/sse"
-    }
-  }
-}
-
-3. 流式 HTTP:
-{
-  "mcpServers": {
-    "http-server": {
-      "url": "https://example.com/mcp"
-    }
-  }
-}`}
-                            {...field}
-                            onChange={(event) => {
-                              field.onChange(event);
-                              setJsonInput(event.target.value);
-                            }}
-                          />
-                        </FormControl>
-                        <FormMessage />
-                      </FormItem>
-                    )}
-                  />
-                </div>
-                <DialogFooter className="mt-4">
-                  <DialogClose asChild>
-                    <Button variant="outline" disabled={isLoading}>
-                      取消
-                    </Button>
-                  </DialogClose>
-                  <Button type="submit" disabled={isLoading}>
-                    {isLoading ? "保存中..." : "保存"}
-                  </Button>
-                </DialogFooter>
-              </form>
-            </Form>
-          </TabsContent>
-        </Tabs>
+        {editor.renderTabs()}
       </DialogContent>
     </Dialog>
   );
