+++ apps/frontend/src/constants/mcp-config-examples.ts
@@ -1,0 +1,35 @@
+/**
+ * MCP 服务配置示例文本
+ * 用于表单高级模式的 placeholder 提示
+ */
+
+export const MCP_CONFIG_PLACEHOLDER_TEXT = `支持三种通信方式：
+
+1. 本地进程 (stdio):
+{
+  "mcpServers": {
+    "local-server": {
+      "command": "npx",
+      "args": ["-y", "@example/mcp-server"]
+    }
+  }
+}
+
+2. 服务器推送 (SSE):
+{
+  "mcpServers": {
+    "sse-server": {
+      "type": "sse",
+      "url": "https://example.com/sse"
+    }
+  }
+}
+
+3. 流式 HTTP:
+{
+  "mcpServers": {
+    "http-server": {
+      "url": "https://example.com/mcp"
+    }
+  }
+}`;
